web/requirements.txt
streamlit
requests
pandas
plotly
python-dotenv
paho-mqtt